Add a request level spec for admin testing
This operates more at the API level, but it appears that Rails5.1 has
taken away access to the session variable, so it becomes necessary to
interact with multiple different controller endpoints, instead of simply
preconfiguring a value in the session.

Apparently, DHH thinks this is an "implementation detail", and that it's
better to require hitting multiple endpoints for any single test (Or,
authenticate through header values).  The man does not care about test
efficency, or any of the myriad corner cases in a typical session-based
app, in the slightest.

https://github.com/rails/rails/issues/23386