reqs for displaying failures dictate we test for a string
we weren't testing for it with the Ruby & Python runners.
the assertion we were using with the Java runner was
imprecise. asserting that it's a string covers "not nil" anyway.
the assertion we were using with the C++ runner also imprecise, but
differently so: "doesn't equal problem.output" doesn't guarantee it's
a string, and in the odd-but-possible situation where it does spit out
the problem output, we honestly don't care--we just need it to be a string.